Move APIModel to separate file
